pyproject.toml
[tool.poetry]
name = "dextro"
version = "0.1.0"
description = ""
authors = ["Kristian Klemon <kristian.klemon@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
dextro = 'dextro.console.entrypoint:main'

[tool.poetry.dependencies]
python = "^3.10"
polars = "^0.20.10"
tqdm = "^4.66.2"
pyarrow = "^15.0.0"
torch = {version = "^2.2.0", optional = true}
lingua-language-detector = {version = "^2.0.2", optional = true}

[tool.poetry.extras]
language-detection = ["lingua-language-detector"]
torch = ["torch"]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
pytest = "^8.0.2"
lorem = "^0.1.1"
pre-commit = "^3.6.2"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
ignore = ["F811"]